回滚功能的基本实现

+++ Concerto_Backend/src/main/java/com/example/concerto/service/TaskVersionService.java
@@ -13,5 +13,5 @@ import java.util.List;
 
 public interface TaskVersionService {
     public List<TaskVersionInfoVo> getTaskVersionInfo(long taskId);
-    public void rollbackVersion(long taskId,long taskversionId);
+    public void rollbackVersion(long taskId,int taskversionId);
 }
